+++ weave/tests/legacy/test_numpy.py
@@ -3,8 +3,8 @@ import numpy as np
 from weave.legacy import artifact_fs, artifact_wandb
 from weave.legacy.ops_domain import table
 
-from ... import types_numpy as numpy_types
 from ... import weave_types as types
+from ...legacy import types_numpy as numpy_types
 
 
 def test_construct_numpy_type():
